// packages/music-core/src/audio/drumOrganismsGenerated.ts
export const GENERATED_DRUM_ORGANISMS: Record<string, unknown> = {
  "swing-default": {
// ...
    "timeSignatureOverrides": {
      "3/4": {
        "verseA": [
          "swing-waltz-12-verse"
        ],
        "bridge": [
          "swing-waltz-12-bridge"
        ],
        "chorus": [
          "swing-waltz-12-chorus"
        ]
      }
    },
// ...
  },
// ...
};